Add files via upload
These are the files for PID control using a Tiva. They work but are only suitable for my specific system. I'm using a 24 VDC JKBLS04 4000rpm BLDC with Hall motor being controlled by a JUKI JYQD-V7.3E3 controller. The controller outputs an aggregate signal from the mix of the Hall outputs and give a 12 pulses per revolution output. This is fed into the Tiva via pin PD6. PWM output is on pin PB6 into the controller to close the loop. This is a video of the system working: https://www.youtube.com/watch?v=WkU8A6H-gFE .Please note that I'm using an Arduino RPM monitor that is taking one Hall output to count RPM. It doesn't work very well and give a bit of a variable output. I'm also using a 12000RPM motor and trying to run at 6250 constant RPM.